Moved the classic Playground class under the legacy package.
