+++ agents_playground/legacy/app/playground_app.py
@@ -4,7 +4,6 @@ import os
 from pathlib import Path
 from typing import Any, List, cast
 
-
 import dearpygui.dearpygui as dpg
 
 from agents_playground.project.create_sim_wizard import CreateSimWizard
