+++ tests/playground_app_test.py
@@ -1,6 +1,6 @@
 import os
 from pathlib import Path
-from agents_playground.app.playground_app import PlaygroundApp
+from agents_playground.legacy.app.playground_app import PlaygroundApp
 from agents_playground.core.observe import Observer
 
 import dearpygui.dearpygui as dpg
